--- networkd/state.go
@@ -1,6 +1,0 @@
-package networkd
-
-type state struct {
-  running bool
-  content map[string]map[string]string
-}
